Auto-calculate dimensions or spacings
Make it easier to switch resolutions for the raster
by specifying, for example, just the desired size
of the raster (extents) and the desired grid size (dimensions)
in terms of the number of voxels.  Alternatively you
can ask for a raster size sampled at a particular
resolution (spacing) and the dimensions are autocalculated.